chore: update e2e tests for android

+++ e2e/helpers/artifacts.js
@@ -1,3 +1,4 @@
+const detox = require('detox');
 const { execSync, spawn, exec } = require('child_process');
 const { existsSync, mkdirSync, unlinkSync } = require('fs');
 const path = require('path');
@@ -18,7 +19,11 @@ const takeScreenshot = () => {
     }
     const screenShotFileName = `${ARTIFACTS_DIR}/screenshot-${screenshotIndex++}.png`;
     try {
-        execSync(`xcrun simctl io booted screenshot ${screenShotFileName}`, SCREENSHOT_OPTIONS);
+        if (detox.device.getPlatform() === 'ios') {
+            execSync(`xcrun simctl io booted screenshot ${screenShotFileName}`, SCREENSHOT_OPTIONS);
+        } else {
+            execSync(`adb shell screencap -p > ${screenShotFileName}`, SCREENSHOT_OPTIONS);
+        }
     } catch (error) {
         console.error('error');
     }
@@ -35,20 +40,25 @@ const startRecordingVideo = () => {
     }
 
     try {
-        spawn('xcrun', ['simctl', 'io', 'booted', 'recordVideo', `${recordingFileName}`], {
-            timeout: 30 * 60 * 1000,
-            maxBuffer: 1024 * 20 * 100,
-        });
+        // video recording only works on ios for now
+        if (detox.device.getPlatform() === 'ios') {
+            spawn('xcrun', ['simctl', 'io', 'booted', 'recordVideo', `${recordingFileName}`], {
+                timeout: 30 * 60 * 1000,
+                maxBuffer: 1024 * 20 * 100,
+            });
+        }
     } catch (error) {
         console.error('error');
     }
 };
 
 const stopRecordingVideo = () => {
-    exec('killall -SIGINT simctl', {
-        timeout: 15 * 1000,
-        maxBuffer: 1024 * 20 * 100,
-    });
+    if (detox.device.getPlatform() === 'ios') {
+        exec('killall -SIGINT simctl', {
+            timeout: 15 * 1000,
+            maxBuffer: 1024 * 20 * 100,
+        });
+    }
 };
 
 module.exports = { takeScreenshot, startRecordingVideo, stopRecordingVideo };
